Fixed no tag read return, changed timed callbacks
modio_read now returns the correct value of 0 if the tag doesn't exist
instead of continuing to try to extract data from the stack.

Timed functions changed so the actual functions can be called externally
and make use of optional parameters. Actual timed callbacks are named
the same but prefixed with a double underscore.