Safety update to read function
Added a length parameter to modio_read that appears before the data
parameter (just to be consistent with the parameter order of
modio_push). This parameter indicates the size of the data array being
passed in and will ensure no out of bounds writing is performed since
crashdetect doesn't seem to pick up out of bounds writing on arrays
which are parameters to a function.

Added a compile flag that enables/disables the warning print that
appears if a tag is not found when trying to read it.